partway through addressing pr feedback.
